Split storage into separate files.
